add dytpe

+++ ezlocalai/VLM.py
@@ -29,7 +29,7 @@ class VLM:
             if torch.cuda.is_available():
                 self.vl_gpt = self.vl_gpt.to(torch.bfloat16).cuda().eval()
             else:
-                self.vl_gpt = self.vl_gpt.eval()
+                self.vl_gpt = self.vl_gpt.to(torch.bfloat16).eval()
         except Exception as e:
             print(f"[VLM] Error: {e}")
             self.vl_chat_processor = None
